pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "bilt"
version = "0.1.0"
description = "BILT (Because I Like Twice) - A PyTorch-based object detection library"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "BILT Contributors"}
]
keywords = ["deep-learning", "object-detection", "pytorch", "computer-vision", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=1.10.0",
    "torchvision>=0.11.0",
    "Pillow>=8.0.0",
    "numpy>=1.19.0",
    "pyyaml>=5.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.9",
]
viz = [
    "matplotlib>=3.3.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/bilt"
Repository = "https://github.com/yourusername/bilt"
Documentation = "https://github.com/yourusername/bilt"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=bilt --cov-report=term-missing"